Add hip and root trajectory drawing. Add shaded bone visualization. Add optional rescaling functionality to the builder. Make hot-reloading a runtime switchable option.

+++ linear_math/quaternion.h
@@ -195,7 +195,7 @@ namespace Math
     quat Result = {};
 
     // Compute the cosine of the angle between the two vectors.
-    double Dot = Math::Dot(Q0.V, Q1.V) + Q0.S*Q1.S;
+    float Dot = Math::Dot(Q0.V, Q1.V) + Q0.S*Q1.S;
 
     // If the dot product is negative, slerp won't take
     // the shorter path. Note that v1 and -v1 are equivalent when
